* ShellCommandReceive code moved to shellcmd_receive.py * ShellCommandReceive class split into multiple platform dependent classes * some other minor changes
